Add Operation defaults trait
